detsys_installer in its own module
