+++ crates/nix_rs/src/env.rs
@@ -3,11 +3,10 @@ use std::{fmt::Display, path::Path};
 
 use bytesize::ByteSize;
 use os_info;
-use regex::Regex;
 use serde::{Deserialize, Serialize};
 use std::process::Command;
-use thiserror::Error;
 use tracing::instrument;
+use crate::detsys_installer::{ NixInstallerVersion, BadInstallerVersion };
 
 /// The environment in which Nix operates
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -219,53 +218,6 @@ impl Display for NixInstaller {
             }
             NixInstaller::Other => write!(f, "Unknown installer"),
         }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, PartialOrd, Eq, Clone)]
-pub struct NixInstallerVersion {
-    pub major: u32,
-    pub minor: u32,
-    pub patch: u32,
-}
-
-impl Display for NixInstallerVersion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}.{}.{}", self.major, self.minor, self.patch)
-    }
-}
-
-#[derive(Error, Debug, Clone, PartialEq)]
-pub enum BadInstallerVersion {
-    #[error("Regex error: {0}")]
-    Regex(#[from] regex::Error),
-    #[error("Failed to decode installer output: {0}")]
-    Decode(#[from] std::string::FromUtf8Error),
-    #[error("Failed to parse installer version: {0}")]
-    Parse(#[from] std::num::ParseIntError),
-    #[error("Failed to fetch installer version")]
-    Command,
-}
-
-impl NixInstallerVersion {
-    pub fn get_version(executable_path: &Path) -> Result<Self, BadInstallerVersion> {
-        let output = std::process::Command::new(executable_path)
-            .arg("--version")
-            .output()
-            .map_err(|_| BadInstallerVersion::Command)?;
-        let output = String::from_utf8(output.stdout)?;
-        let re = Regex::new(r"(?:nix-installer )?(\d+)\.(\d+)\.(\d+)")?;
-
-        let captures = re.captures(&output).ok_or(BadInstallerVersion::Command)?;
-        let major = captures[1].parse::<u32>()?;
-        let minor = captures[2].parse::<u32>()?;
-        let patch = captures[3].parse::<u32>()?;
-
-        Ok(NixInstallerVersion {
-            major,
-            minor,
-            patch,
-        })
     }
 }
